Split local rules into modules and describe every rule
